+++ linqex/__init__.py
@@ -1,2 +1,4 @@
-from linqex.linq import *
 from linqex.build import *
+from linqex.linq import *
+
+version = "1.6"