+++ linqex/linq/__init__.py
@@ -1,6 +1,3 @@
-from linqex._typing import _TK, _TV, _TK2, _TV2, _Key, _Value, _Desc
-from linqex.linq.iterlist import EnumerableList
-from linqex.linq.iterdict import EnumerableDict
-from linqex.linq.iterable import Enumerable
+from linqex.linq.iterable import AbstractEnumerable, Enumerable, EnumerableList, EnumerableItem, EnumerableDict
 
-__all__ = ["_TK", "_TV", "_TK2", "_TV2", "_Key", "_Value", "_Desc", "EnumerableList", "EnumerableDict", "Enumerable"]
+__all__ = ["AbstractEnumerable", "EnumerableList", "EnumerableItem", "EnumerableDict", "Enumerable"]